requirements.txt
dj-database-url
Django
django-filter
django-debug-toolbar
djangorestframework
djangorestframework-simplejwt
python-decouple
